requirements.txt
alembic==1.7.3
click==8.0.1
dnspython==2.1.0
dynaconf==3.1.7
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
Pillow==8.3.2
python-slugify==5.0.2
SQLAlchemy==1.4.23
text-unidecode==1.3
Unidecode==1.3.2
Werkzeug==2.0.1
WTForms==2.3.3
